requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
absl-py==2.2.2
    # via tensorboard
alembic==1.15.2
    # via optuna
bottle==0.13.2
    # via optuna-dashboard
colorlog==6.9.0
    # via optuna
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.18.0
    # via torch
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.71.0
    # via tensorboard
jinja2==3.1.6
    # via torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
mako==1.3.9
    # via alembic
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.10.1
    # via finalproject (pyproject.toml)
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.4
    # via
    #   finalproject (pyproject.toml)
    #   contourpy
    #   matplotlib
    #   optuna
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
optuna==4.2.1
    # via
    #   finalproject (pyproject.toml)
    #   optuna-dashboard
optuna-dashboard==0.18.0
    # via finalproject (pyproject.toml)
packaging==24.2
    # via
    #   matplotlib
    #   optuna
    #   optuna-dashboard
    #   tensorboard
pillow==11.1.0
    # via
    #   matplotlib
    #   torchvision
protobuf==6.30.2
    # via tensorboard
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.2
    # via optuna
scikit-learn==1.6.1
    # via
    #   finalproject (pyproject.toml)
    #   optuna-dashboard
scipy==1.15.2
    # via scikit-learn
setuptools==78.1.0
    # via
    #   tensorboard
    #   torch
six==1.17.0
    # via
    #   python-dateutil
    #   tensorboard
sqlalchemy==2.0.40
    # via
    #   alembic
    #   optuna
standard-imghdr==3.13.0
    # via finalproject (pyproject.toml)
sympy==1.13.1
    # via torch
tensorboard==2.19.0
    # via finalproject (pyproject.toml)
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.6.0
    # via
    #   finalproject (pyproject.toml)
    #   torchvision
torchvision==0.21.0
    # via finalproject (pyproject.toml)
tqdm==4.67.1
    # via optuna
triton==3.2.0
    # via torch
typing-extensions==4.13.1
    # via
    #   alembic
    #   sqlalchemy
    #   torch
werkzeug==3.1.3
    # via tensorboard
